fix: update Users DTOs phone field to be a long instead of int

+++ src/main/java/com/sip/api/dtos/user/UserDto.java
@@ -6,10 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.Set;
-
 
 @Data
 @Builder
@@ -22,7 +20,7 @@ public class UserDto {
     private String firstName;
     private String lastName;
     private LocalDate birthDate;
-    private BigInteger phone;
+    private Long phone;
     private UserStatus status;
     private Set<String> roles;
 }
